support bigger int values for amouont. Improve report: display signs with category

+++ pkg/expense/expenses.go
@@ -14,7 +14,7 @@ const (
 type Expense struct {
 	Date        time.Time
 	Description string
-	Amount      uint16
+	Amount      uint32
 	Decimal     uint16
 	Type        ExpenseType
 	Currency    string
